producer/requirements.txt
kafka-python>=2.0.2
faker>=18.0.0
